refactor: reorder function sequence to make it more reasonable
